Update eos_eapi.py
This commit remove the use of the singular 'vrf' and adds the use of 'vrfs'

The singular VRF posed a few issues.  First is was impossible to define more than one VRF without creating multiple tasks.  This in turn made the module not idempotent.  This reconfiguration of the module allows for none ('default') or multiple VRFs to be configured.  Before, if multiple VRFs were defined and configured in the management API section of the configuration the return of the command 'show management api http-commands | json' has two vrf keys.  One is vrf, the other is vrfs.  vrf appear to alphabetize the configured VRFs and return only the first result.  vrfs returns all configured vrfs.

This module assumes that if the VRF is listed, it should be configured and 'no shutdown' applied as the only options are to 'no shutdown' or 'shutdown' the VRF in the management API configuration.  The module also now removes any VRFs from the management api configuration which are not listed in the playbook.

Additional work would have to be performed in order to add the ability to create comments or access-lists.