Fix idempotency in logging_global module (#575)
* Add unit tests for issue #574

* Update logging_global RM and argspec for #574

We should always add the port to the outgoing command for loggin hosts,
and always add it to the configuration dictionary when parsing config
to ensure that we are idempotent when rendering logging config.

* Update integration tests for #574

Add new logging hosts that omit port and update fixtures/tests to
reflect changes in module behavior.

* Update docstring to reflect default port

* Add changelog fragment for logging_global port

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: Ruchi Pakhle <72685035+Ruchip16@users.noreply.github.com>
Co-authored-by: Sagar Paul <sagpaul@redhat.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>